+++ programs/revenue-distribution/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-- withdraw deposited SOL (#111)
 - register rewards integrations (#113)
+
+## [v0.3.4]
+- withdraw deposited SOL (#111)
 
 ## [v0.3.3]